moved 'slr_data' data files to local storage (where other data are stored) rather than storing on github, and moved check_table_exists() from tide_query_location.py to main_tide_slr.py because it was causing circular error
