+++ src/dynamic_boundary_conditions/main_tide_slr.py
@@ -7,6 +7,7 @@
 import pathlib
 from typing import Union
 
+import sqlalchemy
 import geopandas as gpd
 from shapely.geometry import box
 
@@ -42,6 +43,22 @@ def get_catchment_area(
     return catchment_area
 
 
+def check_table_exists(engine, db_table_name: str) -> bool:
+    """
+    Check if table exists in the database.
+
+    Parameters
+    ----------
+    engine
+        Engine used to connect to the database.
+    db_table_name : str
+        Database table name.
+    """
+    insp = sqlalchemy.inspect(engine)
+    table_exists = insp.has_table(db_table_name, schema="public")
+    return table_exists
+
+
 def main():
     # Connect to the database
     engine = setup_environment.get_database()
@@ -63,9 +80,12 @@ def main():
         tide_length_mins=2880,
         interval_mins=10)
 
-    # Store sea level rise data to database and get closest sea level rise site data from database
-    sea_level_rise_data.store_slr_data_to_db(engine)
+    # Store sea level rise data to database
+    slr_data_dir = config.get_env_variable("DATA_DIR_SLR", cast_to=pathlib.Path)
+    sea_level_rise_data.store_slr_data_to_db(engine, slr_data_dir)
+    # Get closest sea level rise site data from database
     slr_data = sea_level_rise_data.get_closest_slr_data(engine, tide_data_king)
+
     # Combine tide and sea level rise data
     tide_slr_data = tide_slr_combine.get_combined_tide_slr_data(
         tide_data=tide_data_king,
@@ -75,6 +95,7 @@ def main():
         ssp_scenario='SSP1-2.6',
         add_vlm=False,
         percentile=50)
+
     # Generate the model input for BG-Flood
     bg_flood_dir = config.get_env_variable("FLOOD_MODEL_DIR", cast_to=pathlib.Path)
     tide_slr_model_input.generate_uniform_boundary_input(bg_flood_dir, tide_slr_data)
